project/todo-backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
requests
python-multipart
sqlalchemy
psycopg2-binary
nats-py